terraform/cloud_functions/buddies_list_pending/requirements.txt
google-cloud-bigquery
PyJWT
flask
